refactor(ui): split scale logic from preset config
