+++ src/ui.rs
@@ -8,7 +8,6 @@ mod command_palette;
 pub mod components;
 mod controls;
 pub mod data;
-mod density;
 mod fonts;
 mod global_actions;
 mod header;
@@ -19,9 +18,11 @@ pub mod models;
 pub(crate) mod presets;
 mod queue;
 mod right_sidebar;
+mod scale;
 mod scroll_follow;
 mod search;
 mod settings;
 pub mod styling;
 mod troubleshooting;
+mod ui_preset;
 pub mod util;
