+++ src/ui/components/label.rs
@@ -6,7 +6,7 @@ use gpui::{
 use smallvec::SmallVec;
 
 use crate::ui::{
-    density::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
+    scale::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
     styling::{ActiveTheme, StyledExt},
 };
