+++ src/ui/components/menu.rs
@@ -2,7 +2,7 @@ use gpui::{prelude::FluentBuilder, *};
 
 use crate::ui::{
     components::icons::{CHECK, LOCK, icon},
-    density::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
+    scale::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
     styling::theme::Theme,
 };
