+++ src/ui/components/segmented_control.rs
@@ -7,7 +7,7 @@ use gpui::{
 use smallvec::SmallVec;
 
 use crate::ui::{
-    density::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
+    scale::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
     styling::theme::Theme,
 };
