+++ src/ui/controls/replaygain.rs
@@ -12,7 +12,7 @@ use cntp_i18n::tr;
 use gpui::{prelude::FluentBuilder, *};
 
 use crate::ui::{
-    density::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
+    scale::{TextStyle, active_density, active_typography, apply_text_style, scale_px},
     styling::theme::Theme,
 };
