+++ src/ui/header.rs
@@ -18,8 +18,8 @@ use crate::{
             menu_bar::MenuBar,
             window_header::header,
         },
-        density::active_typography,
         library::nav_buttons::nav_buttons,
+        scale::active_typography,
         styling::ActiveTheme,
     },
 };
