+++ src/ui/lyrics.rs
@@ -11,8 +11,8 @@ use crate::{
             icons::{MICROPHONE, icon},
             scrollbar::{RightPad, ScrollableHandle, floating_scrollbar},
         },
-        density::{TextStyle, active_density, interpolate_text_style, scale_px},
         models::{CurrentTrack, Models, PlaybackInfo},
+        scale::{TextStyle, active_density, interpolate_text_style, scale_px},
         scroll_follow::{SmoothScrollFollow, ease_out_cubic},
         styling::ActiveTheme,
     },
