+++ src/ui/ui_preset.rs
@@ -1,4 +1,4 @@
-use gpui::{App, Global, Styled, px};
+use gpui::{App, Global};
 use serde::{
     Deserialize, Serialize,
     de::{MapAccess, Visitor, value::MapAccessDeserializer},
@@ -7,112 +7,13 @@ use serde::{
 use crate::{
     settings::{
         SettingsGlobal,
-        interface::{InterfaceSettings, UiDensity, UiPresetKind, classify_ui_preset_id},
+        interface::{InterfaceSettings, UiPresetKind, classify_ui_preset_id},
     },
     ui::layout::{
         defaults::{default_shell_layout, stage_shell_layout},
         schema::ShellLayout,
     },
 };
-
-#[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize)]
-pub struct TextStyle {
-    pub size: f32,
-    pub line_height: f32,
-}
-
-impl TextStyle {
-    pub const fn new(size: f32, line_height: f32) -> Self {
-        Self { size, line_height }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct TypographyRoles {
-    pub body: TextStyle,
-    pub secondary_body: TextStyle,
-    pub caption: TextStyle,
-    pub label: TextStyle,
-    pub section_title: TextStyle,
-    pub panel_title: TextStyle,
-    pub mono_body: TextStyle,
-}
-
-pub fn interpolate_text_style(
-    density: UiDensity,
-    compact: TextStyle,
-    default: TextStyle,
-    comfortable: TextStyle,
-) -> TextStyle {
-    TextStyle::new(
-        interpolate_scalar(density, compact.size, default.size, comfortable.size),
-        interpolate_scalar(
-            density,
-            compact.line_height,
-            default.line_height,
-            comfortable.line_height,
-        ),
-    )
-}
-
-pub fn interpolate_scalar(density: UiDensity, compact: f32, default: f32, comfortable: f32) -> f32 {
-    density.interpolate(compact, default, comfortable)
-}
-
-pub fn scale_px(density: UiDensity, default: f32, delta: f32) -> f32 {
-    default + (density.value() * delta)
-}
-
-pub fn typography_roles(density: UiDensity) -> TypographyRoles {
-    TypographyRoles {
-        body: interpolate_text_style(
-            density,
-            TextStyle::new(13.0, 17.0),
-            TextStyle::new(14.0, 18.0),
-            TextStyle::new(15.0, 20.0),
-        ),
-        secondary_body: interpolate_text_style(
-            density,
-            TextStyle::new(12.0, 16.0),
-            TextStyle::new(13.0, 17.0),
-            TextStyle::new(14.0, 18.0),
-        ),
-        caption: interpolate_text_style(
-            density,
-            TextStyle::new(11.0, 15.0),
-            TextStyle::new(12.0, 16.0),
-            TextStyle::new(13.0, 18.0),
-        ),
-        label: interpolate_text_style(
-            density,
-            TextStyle::new(13.0, 17.0),
-            TextStyle::new(14.0, 18.0),
-            TextStyle::new(15.0, 20.0),
-        ),
-        section_title: interpolate_text_style(
-            density,
-            TextStyle::new(17.0, 21.0),
-            TextStyle::new(18.0, 22.0),
-            TextStyle::new(20.0, 24.0),
-        ),
-        panel_title: interpolate_text_style(
-            density,
-            TextStyle::new(20.0, 24.0),
-            TextStyle::new(22.0, 26.0),
-            TextStyle::new(24.0, 28.0),
-        ),
-        mono_body: interpolate_text_style(
-            density,
-            TextStyle::new(13.0, 17.0),
-            TextStyle::new(14.0, 18.0),
-            TextStyle::new(15.0, 20.0),
-        ),
-    }
-}
-
-pub fn active_typography(cx: &App) -> TypographyRoles {
-    typography_roles(active_density(cx))
-}
 
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct FlatOptionalLayout(pub Option<ShellLayout>);
@@ -275,10 +176,6 @@ pub struct UiPresetConfigGlobal(pub UiPresetConfig);
 
 impl Global for UiPresetConfigGlobal {}
 
-pub fn resolve_density(interface: &InterfaceSettings) -> UiDensity {
-    interface.ui_density
-}
-
 pub fn resolve_shell_layout(
     interface: &InterfaceSettings,
     preset: Option<&UiPresetConfig>,
@@ -292,16 +189,6 @@ pub fn resolve_shell_layout(
     }
 }
 
-pub fn active_density(cx: &App) -> UiDensity {
-    let interface = cx
-        .global::<SettingsGlobal>()
-        .model
-        .read(cx)
-        .interface
-        .clone();
-    resolve_density(&interface)
-}
-
 pub fn active_shell_layout(cx: &App) -> ShellLayout {
     let interface = cx
         .global::<SettingsGlobal>()
@@ -313,21 +200,14 @@ pub fn active_shell_layout(cx: &App) -> ShellLayout {
     resolve_shell_layout(&interface, Some(&preset))
 }
 
-pub fn apply_text_style<T>(dest: T, style: TextStyle) -> T
-where
-    T: Styled,
-{
-    dest.text_size(px(style.size))
-        .line_height(px(style.line_height))
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::ui::layout::defaults::default_shell_layout;
+    use crate::{
+        settings::interface::UiDensity,
+        ui::layout::defaults::{default_shell_layout, stage_shell_layout},
+    };
 
-    use crate::settings::interface::UiDensity;
-
-    use super::{TextStyle, UiPresetConfig, resolve_density, resolve_shell_layout, scale_px};
+    use super::{UiPresetConfig, resolve_shell_layout};
 
     fn interface(
         ui_preset: Option<&str>,
@@ -341,52 +221,19 @@ mod tests {
     }
 
     #[test]
-    fn density_comes_from_settings() {
-        let resolved = resolve_density(&interface(
-            Some("layouts/custom.ron"),
-            UiDensity::COMFORTABLE,
-        ));
-
-        assert_eq!(resolved, UiDensity::COMFORTABLE);
-    }
-
-    #[test]
-    fn text_style_interpolates_between_anchors() {
-        assert_eq!(
-            super::interpolate_text_style(
-                UiDensity::from(-0.5),
-                TextStyle::new(12.0, 16.0),
-                TextStyle::new(14.0, 18.0),
-                TextStyle::new(16.0, 20.0),
-            ),
-            TextStyle::new(13.0, 17.0)
-        );
-
-        assert_eq!(
-            super::interpolate_text_style(
-                UiDensity::from(0.5),
-                TextStyle::new(12.0, 16.0),
-                TextStyle::new(14.0, 18.0),
-                TextStyle::new(16.0, 22.0),
-            ),
-            TextStyle::new(15.0, 20.0)
-        );
-    }
-
-    #[test]
-    fn scale_px_offsets_from_default() {
-        assert_eq!(scale_px(UiDensity::COMPACT, 36.0, 2.0), 34.0);
-        assert_eq!(scale_px(UiDensity::DEFAULT, 36.0, 2.0), 36.0);
-        assert_eq!(scale_px(UiDensity::COMFORTABLE, 36.0, 2.0), 38.0);
-    }
-
-    #[test]
-    fn custom_layout_falls_back_to_default_shell_layout_when_missing() {
+    fn file_preset_falls_back_to_default_shell_layout_when_missing() {
         let resolved = resolve_shell_layout(
             &interface(Some("layouts/custom.ron"), UiDensity::DEFAULT),
             Some(&UiPresetConfig::default()),
         );
 
         assert_eq!(resolved, default_shell_layout());
+    }
+
+    #[test]
+    fn stage_preset_uses_builtin_stage_shell_layout() {
+        let resolved = resolve_shell_layout(&interface(Some("stage"), UiDensity::DEFAULT), None);
+
+        assert_eq!(resolved, stage_shell_layout());
     }
 }
